+++ src/abstraction/plugin.options.ts
@@ -1,7 +1,12 @@
 import Vue from "vue";
 export interface PluginOptions  {
-    debug: boolean;
-    injections: {
+    debug?: boolean;
+    externals?: {
         [moduleName: string]: { [componentName: string]: Vue}
+    },
+    components?: {
+        entry?: string | string []
+        deep?: boolean;
+        path?: string | string []
     }
 }