+++ src/classification/dependency.class.ts
@@ -2,13 +2,14 @@ import {posix} from "path";
 import {normalize} from "upath";
 import { Argumenter } from "@joejukan/argumenter";
 
-const REGEX_VUE = /^[\w\_\-\.]*\.vue$/i;
+const REGEX_VUE = /\.vue$/i;
 export class DependencyClass {
     public name: string;
     public path: string;
     public symbol: string;
     public defaulted: boolean;
 
+    public constructor();
     public constructor(name: string);
     public constructor(name: string, defaulted: boolean);
     public constructor(name: string, symbol: string);