+++ src/function/pitch.ts
@@ -1,7 +1,9 @@
-import {ASTClass} from "../classification";
+import {ASTClass, VueDIPlugin} from "../classification";
 import {dependencies} from "../globalization";
 
 export function pitch(this: Context, remainingPath, precedingPath, data: {[key: string]: any}){
-    let ast = new ASTClass();
-    ast.pitch(remainingPath);
+    if(!VueDIPlugin.pitched && /\.vue$/i.test(remainingPath)){
+        let ast = new ASTClass();
+        ast.pitch(remainingPath);
+    }
 }